Modularize Firebase upload functionality
Create new class UploadToFireBase to handle data uploads.

- Introduce UploadToFireBase class to encapsulate Firebase operations
- Move upload logic for both storage and Realtime Database to new class
- Refactor AddItemFragment to utilize UploadToFireBase for data submission
- Improve code organization and separation of concerns
- Enhance maintainability and reusability of Firebase upload operations

This change modularizes the code responsible for uploading data to
Firebase Storage and Realtime Database, promoting better code structure
and easier future maintenance.
